+++ src/test/java/domain/participant/PlayersTest.java
@@ -4,7 +4,6 @@ import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.Assertions.assertThatNoException;
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
 
-import domain.BettingMoney;
 import domain.card.Card;
 import domain.card.DrawnCards;
 import java.util.ArrayList;
